requirements.txt
streamlit
pandas
plotly
pyarrow
Pillow
numpy
